+++ src/50expression.js
@@ -262,7 +262,7 @@ yy.Op.prototype.toType = function(tableid) {
 		}
 	}
 	
-	if(['AND','OR','NOT','=','==','===', '!=','!==','!===','>','>=','<','<=', 'IN', 'NOT IN', 'LIKE', 'NOT LIKE'].indexOf(this.op) >-1 ){
+	if(['AND','OR','NOT','=','==','===', '!=','!==','!===','>','>=','<','<=', 'IN', 'NOT IN', 'LIKE', 'NOT LIKE', 'REGEXP'].indexOf(this.op) >-1 ){
 		return 'boolean';
 	}
 
@@ -395,6 +395,13 @@ yy.Op.prototype.toJS = function(context,tableid,defcols) {
 		}
 		s += '))';
 		return s;
+	}
+	if(this.op === 'REGEXP') {
+		return 'alasql.stdfn.REGEXP_LIKE(' 
+			+ leftJS()
+			+ ','
+			+ rightJS()
+			+ ')';
 	}
 
 	if(this.op === 'BETWEEN' || this.op === 'NOT BETWEEN') {