+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Added SUBSTR() alias for MID() function (for SQLite compatibility)
 * Changed LIKE function to alasql.utils.like
 * Added LIKE ESCAPE functionality
+* Added REGEXP operator (like MySQL) and REGEXP_LIKE() function (like in Oracle)
 
 ## Version 0.2.0 "Athens" (13.07.2015)
 The purpose of this release were hard work on: