+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Added LIKE ESCAPE functionality
 * Added REGEXP operator (like MySQL) and REGEXP_LIKE() function (like in Oracle)
 * INSERT OR REPLACE VALUE, INSERT OR REPLACE SELECT
+* Read Blob as parameter for from-functions like XLS()
 
 ## Version 0.2.0 "Athens" (13.07.2015)
 The purpose of this release were hard work on: