+++ CHANGELOG.md
@@ -1,14 +1,16 @@
 # Changelog
-
-### 0.2.1 "Rodos" (13.07.2015 - ...)
-
 * Check https://trello.com/b/qxz65pVi/alasql-roadmap for roadmap
+
+### 0.2.1 "Rodos" (xx.09.2015)
 * Added SUBSTR() alias for MID() function (for SQLite compatibility)
 * Changed LIKE function to alasql.utils.like
 * Added LIKE ESCAPE functionality
 * Added REGEXP operator (like MySQL) and REGEXP_LIKE() function (like in Oracle)
 * INSERT OR REPLACE VALUE, INSERT OR REPLACE SELECT
 * Read Blob as parameter for from-functions like XLS()
+* Bugfix related to 'NOT' and '=' predecession
+* Added FETCH NEXT
+* Bugfix alasql for running in cordova on iOS
 
 ## Version 0.2.0 "Athens" (13.07.2015)
 The purpose of this release were hard work on: